add: separet goodwallet/userstorage instance import from class
